+++ StudyLink/src/controller/AccountManagement.java
@@ -1,4 +1,4 @@
-package view;
+package controller;
 
 import view.dashboard.Dashboard;
