+++ StudyLink/src/controller/ForgotPassword.java
@@ -1,4 +1,4 @@
-package view;
+package controller;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
