+++ StudyLink/src/controller/Main.java
@@ -2,12 +2,10 @@ package controller;
 
 import java.sql.*;
 
-import view.Login;
-
 import javax.swing.*;
 
 public class Main extends JFrame {
-    
+
     public static void main(String[] args) throws SQLException {
         SwingUtilities.invokeLater(new Runnable() {
             @Override
