+++ StudyLink/src/controller/UploadFile.java
@@ -1,4 +1,4 @@
-package view;
+package controller;
 import helpers.*;
 import model.Course;
