+++ StudyLink/src/view/dashboard/Dashboard.java
@@ -4,7 +4,9 @@ import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
 
-import view.*;
+import controller.AccountManagement;
+import controller.Login;
+import controller.UploadFile;
 
 public class Dashboard implements ActionListener, DashboardGUI{
